+++ PSUtil/internal/TEPP/prompt.Tepp.ps1
@@ -1,0 +1,9 @@
+﻿Register-PSFTeppScriptblock -Name 'psutil.prompt' -ScriptBlock {
+	$module = Get-Module PSUtil
+	& $module {
+		foreach ($item in (Get-ChildItem "$script:ModuleRoot\internal\prompts"))
+		{
+			$item.BaseName -replace '\.prompt$'
+		}
+	}
+}